move all agg / window function from dir colexec to function (#11775)
decoupling agg/win function's execution and implement.
support implement aggregation and window functions in function framework directly.

Approved by: @iamlinjunhong, @zengyan1, @XuPeng-SH, @nnsgmsone

+++ pkg/sql/compile/scopeRemoteRun.go
@@ -1498,7 +1498,7 @@ func convertToPipelineAggregates(ags []agg.Aggregate) []*pipeline.Aggregate {
 	result := make([]*pipeline.Aggregate, len(ags))
 	for i, a := range ags {
 		result[i] = &pipeline.Aggregate{
-			Op:     int32(a.Op),
+			Op:     a.Op,
 			Dist:   a.Dist,
 			Expr:   a.E,
 			Config: a.Config,
@@ -1512,7 +1512,7 @@ func convertToAggregates(ags []*pipeline.Aggregate) []agg.Aggregate {
 	result := make([]agg.Aggregate, len(ags))
 	for i, a := range ags {
 		result[i] = agg.Aggregate{
-			Op:     int(a.Op),
+			Op:     a.Op,
 			Dist:   a.Dist,
 			E:      a.Expr,
 			Config: a.Config,
